Rebase

+++ src/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Burning.java
@@ -3,7 +3,7 @@
  * Copyright (C) 2012-2015  Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2015 Evan Debenham
+ * Copyright (C) 2014-2016 Evan Debenham
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,20 +22,24 @@ package com.shatteredpixel.shatteredpixeldungeon.actors.buffs;
 
 import com.shatteredpixel.shatteredpixeldungeon.Badges;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
-import com.shatteredpixel.shatteredpixeldungeon.ResultDescriptions;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Blob;
 import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Fire;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.Thief;
+import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
 import com.shatteredpixel.shatteredpixeldungeon.effects.particles.ElmoParticle;
 import com.shatteredpixel.shatteredpixeldungeon.items.Heap;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
+import com.shatteredpixel.shatteredpixeldungeon.items.armor.glyphs.Brimstone;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.ChargrilledMeat;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.MysteryMeat;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfElements.Resistance;
 import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.Scroll;
+import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfMagicalInfusion;
+import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfUpgrade;
 import com.shatteredpixel.shatteredpixeldungeon.levels.Level;
+import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.GameScene;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.CharSprite;
 import com.shatteredpixel.shatteredpixeldungeon.ui.BuffIndicator;
@@ -44,9 +48,6 @@ import com.watabou.utils.Bundle;
 import com.watabou.utils.Random;
 
 public class Burning extends Buff implements Hero.Doom {
-
-	private static final String TXT_BURNS_UP		= "%s burns up!";
-	private static final String TXT_BURNED_TO_DEATH	= "You burned to death...";
 	
 	private static final float DURATION = 8f;
 	
@@ -74,38 +75,66 @@ public class Burning extends Buff implements Hero.Doom {
 	public boolean act() {
 		
 		if (target.isAlive()) {
-			
-			target.damage( Random.Int( 1, 5 ), this );
+
+			//maximum damage scales from 6 to 2 depending on remaining hp.
+			int maxDmg = 3 + Math.round( 4 * target.HP / (float)target.HT );
+			int damage = Random.Int( 1, maxDmg );
 			Buff.detach( target, Chill.class);
 
 			if (target instanceof Hero) {
 
 				Hero hero = (Hero)target;
-				Item item = hero.belongings.randomUnequipped();
-				if (item instanceof Scroll) {
-					
-					item = item.detach( hero.belongings.backpack );
-					GLog.w( TXT_BURNS_UP, item.toString() );
-					
-					Heap.burnFX( hero.pos );
-					
-				} else if (item instanceof MysteryMeat) {
-					
-					item = item.detach( hero.belongings.backpack );
-					ChargrilledMeat steak = new ChargrilledMeat();
-					if (!steak.collect( hero.belongings.backpack )) {
-						Dungeon.level.drop( steak, hero.pos ).sprite.drop();
+
+				if (hero.belongings.armor != null && hero.belongings.armor.hasGlyph(Brimstone.class)){
+
+					float heal = hero.belongings.armor.level()/5f;
+					if (Random.Float() < heal % 1) heal++;
+					if (heal >= 1 && hero.HP < hero.HT) {
+						hero.sprite.emitter().burst(Speck.factory(Speck.HEALING), (int)heal);
+						hero.HP = Math.min(hero.HT, hero.HP + (int)heal);
 					}
-					GLog.w( TXT_BURNS_UP, item.toString() );
-					
-					Heap.burnFX( hero.pos );
-					
+
+				} else {
+
+					hero.damage( damage, this );
+					Item item = hero.belongings.randomUnequipped();
+					if (item instanceof Scroll
+							&& !(item instanceof ScrollOfUpgrade || item instanceof ScrollOfMagicalInfusion)) {
+
+						item = item.detach( hero.belongings.backpack );
+						GLog.w( Messages.get(this, "burnsup", Messages.capitalize(item.toString())) );
+
+						Heap.burnFX( hero.pos );
+
+					} else if (item instanceof MysteryMeat) {
+
+						item = item.detach( hero.belongings.backpack );
+						ChargrilledMeat steak = new ChargrilledMeat();
+						if (!steak.collect( hero.belongings.backpack )) {
+							Dungeon.level.drop( steak, hero.pos ).sprite.drop();
+						}
+						GLog.w( Messages.get(this, "burnsup", item.toString()) );
+
+						Heap.burnFX( hero.pos );
+
+					}
+
 				}
 				
-			} else if (target instanceof Thief && ((Thief)target).item instanceof Scroll) {
-				
-				((Thief)target).item = null;
-				target.sprite.emitter().burst( ElmoParticle.FACTORY, 6 );
+			} else {
+				target.damage( damage, this );
+			}
+
+			if (target instanceof Thief) {
+
+				Item item = ((Thief) target).item;
+
+				if (item instanceof Scroll &&
+						!(item instanceof ScrollOfUpgrade || item instanceof ScrollOfMagicalInfusion)) {
+					target.sprite.emitter().burst( ElmoParticle.FACTORY, 6 );
+					((Thief)target).item = null;
+				}
+
 			}
 
 		} else {
@@ -120,7 +149,6 @@ public class Burning extends Buff implements Hero.Doom {
 		left -= TICK;
 		
 		if (left <= 0 ||
-			Random.Float() > (2 + (float)target.HP / target.HT) / 3 ||
 			(Level.water[target.pos] && !target.flying)) {
 			
 			detach();
@@ -145,8 +173,13 @@ public class Burning extends Buff implements Hero.Doom {
 	}
 
 	@Override
+	public String heroMessage() {
+		return Messages.get(this, "heromsg");
+	}
+
+	@Override
 	public String toString() {
-		return "Burning";
+		return Messages.get(this, "name");
 	}
 
 	public static float duration( Char ch ) {
@@ -156,14 +189,7 @@ public class Burning extends Buff implements Hero.Doom {
 
 	@Override
 	public String desc() {
-		return "Few things are more distressing than being engulfed in flames.\n" +
-				"\n" +
-				"Fire will deal damage every turn until it is put out by water, expires, or it is resisted. " +
-				"Fire can be extinquished by stepping into water, or from the splash of a shattering potion. \n" +
-				"\n" +
-				"Additionally, the fire may ignite flammable terrain or items that it comes into contact with.\n" +
-				"\n" +
-				"The burning will last for " + dispTurns(left) + ", or until it is resisted or extinquished.";
+		return Messages.get(this, "desc", dispTurns(left));
 	}
 
 	@Override
@@ -171,7 +197,7 @@ public class Burning extends Buff implements Hero.Doom {
 		
 		Badges.validateDeathFromFire();
 		
-		Dungeon.fail( ResultDescriptions.BURNING );
-		GLog.n( TXT_BURNED_TO_DEATH );
+		Dungeon.fail( getClass() );
+		GLog.n( Messages.get(this, "ondeath") );
 	}
 }
